Implement function invocation in parser
Added ability to parse function calls in the parser by creating a new FunctionNode class and extending the existing expression method. This ensures function invocations are dealt with correctly, updating the Parse to handle parameters and creation of these new Function Nodes. Additionally, adjusted variable and expression casting to align with the introduced changes.